Refactoring project structure to be more feature-based
