+++ App/views/AdminDashboard/index.js
@@ -5,7 +5,7 @@ import {
   Text,
   View
 } from 'react-native';
-import UnassignedEventListContainer from '../../components/EventList/UnassignedEventListContainer.js';
+import UnassignedEventListContainer from '../../components/EventList/container/UnassignedEventList.js';
 
 export default class AdminDashboard extends Component {
   static navigationOptions = {
